Added some shit
Now a new list of items is created called groupedCart which contains one of each element in the cart + how many of each element is in the cart, which makes for a prettier display in the cart overview. Added plus and minus icons to quantity which is not implemented yet. Added a remove button which removes the entire item from the cart.